added ModeCode's fixed syntax & intent
Added several ModeCode's to a switch statement and fixed the spelling of
broadcast_cmd_reserved->broadcast_cmd_received, since the padding for
the reserved bits is added as a later constant. Also, fixed the
sequencing of the actual transmittal of packets onto the bus by the
virtual Bus Controller and avionics.